github_uploader/requirements.txt
Flask
requests
PyGithub